refactor: 메소드 위치 변경 및 사용하지 않는 코드 제거
