+++ src/main/java/bowling/domain/pin/Gutter.java
@@ -5,12 +5,12 @@ public class Gutter extends FramePins {
         super(firstPins, secondPins);
     }
 
+    public static Gutter of() {
+        return new Gutter(Pins.of(0), Pins.of(0));
+    }
+
     @Override
     public boolean isEnd() {
         return true;
-    }
-
-    public static Gutter of() {
-        return new Gutter(Pins.of(0), Pins.of(0));
     }
 }
