+++ src/test/java/bowling/domain/pin/FramePinsCreatorTest.java
@@ -1,5 +1,6 @@
 package bowling.domain.pin;
 
+import bowling.domain.pin.creator.FramePinsCreator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
